[level 2] Title: 시소 짝꿍, Time: 0.02 ms, Memory: 4.14 MB -BaekjoonHub

+++ 프로그래머스/2/152996. 시소 짝꿍/시소 짝꿍.cpp
@@ -7,12 +7,12 @@ using namespace std;
 
 bool check(int a, int b)
 {
-    long long alens[3] = {a * 2, a * 3, a * 4};    
-    long long blens[3] = {b * 2, b * 3, b * 4};
+    int alens[3] = {a * 2, a * 3, a * 4};    
+    int blens[3] = {b * 2, b * 3, b * 4};
     
-    for (long long i = 0; i < 3; ++i)
+    for (int i = 0; i < 3; ++i)
     {
-        for (long long j = 0; j < 3; ++j)
+        for (int j = 0; j < 3; ++j)
         {
             if (alens[i] == blens[j])
                 return true;
@@ -33,8 +33,7 @@ long long solution(vector<int> weights) {
         if (weight_counts[i] == 0)        
             continue;        
         else if (weight_counts[i] > 1)
-            answer += (weight_counts[i] * (weight_counts[i] -1)) / 2;
-        
+            answer += (weight_counts[i] * (weight_counts[i] -1)) / 2;        
         for (int j = i + 1; j < 1001; ++j)
         {
             if (weight_counts[j] == 0)
